ModelEngine/requirements.txt
huggingface-hub>=0.19.0
pyyaml>=6.0.1
aiohttp>=3.9.0
